src/backend/src/requirements.txt
psutil
pillow
tensorflow<2.11
keras_vggface
keras_facenet
opencv-python
numpy
scipy
scikit-learn
dlib
redis
docker
uuid
psycopg2-binary
argparse
pypiwin32

```